GH2883 Move ServicesInfoService to the apiml-common
