+++ gateway-service/src/main/java/org/zowe/apiml/gateway/services/ServerInfoConfig.java
@@ -10,9 +10,13 @@
 
 package org.zowe.apiml.gateway.services;
 
+import com.netflix.discovery.EurekaClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.zowe.apiml.eurekaservice.client.util.EurekaMetadataParser;
+import org.zowe.apiml.product.gateway.GatewayConfigProperties;
+import org.zowe.apiml.product.routing.transform.TransformService;
+import org.zowe.apiml.product.services.ServicesInfoService;
 
 @Configuration
 public class ServerInfoConfig {
@@ -20,6 +24,13 @@ public class ServerInfoConfig {
     @Bean
     public EurekaMetadataParser getEurekaMetadataParser() {
         return new EurekaMetadataParser();
+    }
+
+
+    @Bean
+    public ServicesInfoService servicesInfoService(EurekaClient eurekaClient,
+                                                   EurekaMetadataParser eurekaMetadataParser, GatewayConfigProperties gatewayConfigProperties, TransformService transformService) {
+        return new ServicesInfoService(eurekaClient, eurekaMetadataParser, gatewayConfigProperties, transformService);
     }
 
 }
