+++ gateway-service/src/main/java/org/zowe/apiml/gateway/services/ServicesInfoController.java
@@ -13,14 +13,23 @@ package org.zowe.apiml.gateway.services;
 import com.netflix.appinfo.InstanceInfo;
 import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.*;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.zowe.apiml.product.services.ServiceInfo;
+import org.zowe.apiml.product.services.ServicesInfoService;
 
 import java.util.List;
 
-import static org.zowe.apiml.gateway.services.ServicesInfoService.CURRENT_VERSION;
-import static org.zowe.apiml.gateway.services.ServicesInfoService.VERSION_HEADER;
+import static org.zowe.apiml.product.services.ServicesInfoService.CURRENT_VERSION;
+import static org.zowe.apiml.product.services.ServicesInfoService.VERSION_HEADER;
+
 
 @RestController
 @RequiredArgsConstructor
